Move mockHttpClient() to separate helpers module
